Added delete ability, configuraible user interface

+++ app/code/Setup.php
@@ -26,7 +26,7 @@ class Setup
     entity_id INT(6) UNSIGNED AUTO_INCREMENT PRIMARY KEY,
     comment VARCHAR(30) NOT NULL,
     parent INT(10),
-    user_id INT(20) NOT NULL,
+    user_id BIGINT(30) NOT NULL,
     create_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
     )";
